+++ src/main/java/so/sao/shop/supplier/pojo/output/CommodityExportOutput.java
@@ -1,5 +1,6 @@
 package so.sao.shop.supplier.pojo.output;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
@@ -44,11 +45,11 @@ public class CommodityExportOutput extends BaseResult{
     /**
      * 市场价
      */
-    private Double price;
+    private BigDecimal price;
     /**
      * 售价
      */
-    private Double unitPrice;
+    private BigDecimal unitPrice;
     /**
      * 库存
      */
@@ -133,19 +134,19 @@ public class CommodityExportOutput extends BaseResult{
         this.ruleValue = ruleValue;
     }
 
-    public Double getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(Double price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
 
-    public Double getUnitPrice() {
+    public BigDecimal getUnitPrice() {
         return unitPrice;
     }
 
-    public void setUnitPrice(Double unitPrice) {
+    public void setUnitPrice(BigDecimal unitPrice) {
         this.unitPrice = unitPrice;
     }
